--- context.gohtml
@@ -1,19 +1,0 @@
-{{.Title}}
-{{.HTML}}
-{{.SafeHTML}}
-{{.}}
-
-<a title="{{.Title}}">
-<a title="{{.HTML}}">
-
-<a href="{{.HTML}}">
-<a href="?q={{.HTML}}">
-<a href="{{.Path}}">
-<a href="?q={{.Path}}">
-
-
-<script>
-  var dog = {{.Dog}};
-  var map = {{.Map}};
-  doWork({{.Title}});
-</script>
