+++ display/hello.gohtml
@@ -1,0 +1,1 @@
+<h1>Hello, {{.Name}}!</h1>
