--- hello.gohtml
@@ -1,1 +1,0 @@
-<h1>Hello, {{.Name}}!</h1>
